+++ Lab1Template/Shaders/trail_vert.glsl
@@ -16,7 +16,7 @@ uniform int flag;
 // bone adjustments
 const int MAX_BONES = 100;
 uniform int boneToggle;
-attribute ivec4 boneIds;
+attribute vec4 boneIds;
 attribute vec4 boneWeights;
 uniform mat4 bones[MAX_BONES];
 
@@ -55,10 +55,10 @@ void main()
 
     if (boneToggle == 1)
     {
-        mat4 boneTrans = bones[boneIds[0]] * boneWeights[0];
-        boneTrans += bones[boneIds[1]] * boneWeights[1];
-        boneTrans += bones[boneIds[2]] * boneWeights[2];
-        boneTrans += bones[boneIds[3]] * boneWeights[3];
+        mat4 boneTrans = bones[int(boneIds[0])] * boneWeights[0];
+        boneTrans += bones[int(boneIds[1])] * boneWeights[1];
+        boneTrans += bones[int(boneIds[2])] * boneWeights[2];
+        boneTrans += bones[int(boneIds[3])] * boneWeights[3];
 
         gl_Position = uProjMatrix * uViewMatrix * uModelMatrix * boneTrans * vertPos;
         pos = uViewMatrix * boneTrans * uModelMatrix * vertPos;
